pyproject.toml
[project]
name = "full-stack-template"
version = "0.1.0"
description = "Full stack template with FastAPI, SQLModel, and React for Techletes projects"
readme = "README.md"
requires-python = ">=3.12,<4.0"
authors = [{ name = "Thom Ezechiels", email = "thom@techletes.ai" }]
maintainers = [{ name = "Techletes DevOps Team", email = "dev@techletes.ai" }]
keywords = ["fastapi", "full-stack", "python", "react", "sqlmodel", "template"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.12",
]

dependencies = [
  "alembic>=1.18.4,<2.0.0",
  "boto3>=1.35.0,<2.0.0",
  "chardet>=7.0.0,<8.0.0",
  "email-validator>=2.3.0,<3.0.0",
  "emails>=0.6,<1.0.0",
  "fastapi[standard]>=0.133.0,<1.0.0",
  "httpx>=0.28.1,<1.0.0",
  "jinja2>=3.1.6,<4.0.0",
  "psycopg[binary]>=3.3.3,<4.0.0",
  "pwdlib[argon2,bcrypt]>=0.3.0,<1.0.0",
  "pydantic>=2.12.5,<3.0.0",
  "pydantic-settings>=2.13.1,<3.0.0",
  "pyjwt>=2.11.0,<3.0.0",
  "python-multipart>=0.0.22,<1.0.0",
  "redis>=7.0.0,<8.0.0",
  "rq>=2.0.0,<3.0.0",
  "sentry-sdk[fastapi]>=2.53.0,<3.0.0",
  "sqlmodel>=0.0.37,<1.0.0",
  "tenacity>=9.1.4,<10.0.0",
]

[project.urls]
Homepage = "https://github.com/TECHLETES/full-stack-template"
Issues = "https://github.com/TECHLETES/full-stack-template/issues"
Repository = "https://github.com/TECHLETES/full-stack-template"

[dependency-groups]
dev = [
  "bandit>=1.7.0,<2.0.0",
  "black>=25.12.0,<27.0.0",
  "build>=0.11.0",
  "coverage<8.0.0,>=7.4.3",
  "fakeredis>=2.28.0",
  "genbadge[coverage]>=1.1.0,<2.0.0",
  "isort>=5.13.0,<9.0.0",
  "mypy<2.0.0,>=1.8.0",
  "pre-commit>=4.5.1,<5.0.0",
  "prek>=0.2.24,<1.0.0",
  "pytest<10.0.0,>=9.0.2",
  "pytest-cov>=7.0.0,<8.0.0",
  "pyupgrade>=3.15.0,<4.0.0",
  "ruff<1.0.0,>=0.2.2",
  "testcontainers>=4.14.1",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.bandit]
exclude_dirs = ["tests", ".venv", "alembic"]
skips = ["B101", "B601"]

[tool.black]
line-length = 88
target-version = ["py312"]
exclude = '''
/(
    \.venv
  | venv
  | alembic
  | tests
  | build
  | dist
  | .github
)/
'''

[tool.coverage]
run.source = ["backend"]
run.omit = [
  "*/__pycache__/*",
  "*/tests/*",
  "*/example/*",
  "*/.venv/*",
  "*/venv/*",
  "*/__init__.py",
  # Add any other files you want to exclude from test coverage
]
report.show_missing = true
report.sort = "-Cover"
html.show_contexts = true

[tool.mypy]
python_version = "3.12"

# ignore_missing_imports for third-party libs without type hints
ignore_missing_imports = true
follow_imports = "silent"

# strictness settings - adjust as needed
warn_unused_ignores = true
warn_return_any = true
warn_unreachable = true
no_implicit_optional = true
allow_redefinition = true
show_column_numbers = true
pretty = true
warn_unused_configs = true

# enforce typing on own code
disallow_untyped_defs = true
disallow_incomplete_defs = true
exclude = ["venv", ".venv"]

# Completely ignore test files
[[tool.mypy.overrides]]
module = ["backend.tests.*", "tests.*"]
ignore_errors = true

# Loosen rules for alembic migrations
[[tool.mypy.overrides]]
module = ["backend.alembic.*"]
disallow_untyped_defs = false
disallow_incomplete_defs = false
warn_return_any = false

[[tool.mypy.overrides]]
module = ["backend.alembic.versions.*"]
ignore_errors = true

[tool.pytest]
addopts = [
  "--cov-report=html",
  "--cov-report=term-missing",
  "--cov-report=xml",
  "--strict-config",
  "--strict-markers",
]
minversion = "9.0"
python_classes = ["Test*"]
python_files = ["*_test.py", "test_*.py"]
python_functions = ["test_*"]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 88
fix = true
respect-gitignore = true
exclude = ["alembic", ".github", "tests"]
lint.dummy-variable-rgx = "^_|^__|^pytest$"
lint.select = [
  "E",  # pycodestyle errors
  "W",  # pycodestyle warnings
  "F",  # pyflakes
  "I",  # isort
  "B",  # flake8-bugbear
  "C4",  # flake8-comprehensions
  "UP",  # pyupgrade
  "ARG001",  # unused arguments in functions
  "T201",  # print statements are not allowed
]
lint.ignore = [
  "E501",  # line too long, handled by black
  "B008",  # do not perform function calls in argument defaults
  "W191",  # indentation contains tabs
  "B904",  # Allow raising exceptions without from e, for HTTPException
]

[tool.setuptools.packages.find]
where = ["."]
include = ["backend"]
exclude = ["tests*"]

[tool.uv]
required-version = ">=0.9.25,<1.0.0"
